requirements.txt
Flask==1.0.2
psycopg2-binary==2.7.6.1